+++ accounting/users/__init__.py
@@ -1,0 +1,2 @@
+from .models import User
+from .routing import user_router
